modularize validations working
