+++ Client/src/views/Login/clientLogin.jsx
@@ -1,21 +1,21 @@
 import React, { useState } from 'react'
 import style from './clientLogin.module.scss'
+import { useValidations } from '../../utils/validationutils'
 const ClientLogin = () => {
+  const {errors, validate} = useValidations();
     const [input, setInput] = useState({
         email: '',
         password: '',
     })
 
-    const [errors, setErrors] = useState({
-        email: '',
-        password: '',
-    })
+   
     const handleChange = (event) => {
         console.log(event)
         setInput({
             ...input,
             [event.target.name]: event.target.value,
         })
+        ///validations ///
         validate(
             {
                 ...input,
@@ -23,6 +23,7 @@ const ClientLogin = () => {
             },
             event.target.name
         )
+         ///validations ///
     }
     const handleSubmit = (e) => {
         e.preventDefault()
@@ -33,53 +34,53 @@ const ClientLogin = () => {
         //navigate home / search auxies ///
     }
 
-    ///validations ///
-    const validate = (input, name) => {
-        ///email validations///
-        if (name === 'email') {
-            if (input.email !== '') {
-                // Use a regular expression to check if the input value is a valid email
-                const emailPattern = new RegExp(
-                    /^([A-Za-z0-9_\-\.])+\@([A-Za-z0-9_\-\.])+\.([A-Za-z]{2,4})$/ //eslint-disable-line
-                )
-                if (emailPattern.test(input.email)) {
-                    setErrors({ ...errors, email: '' })
-                } else {
-                    setErrors({ ...errors, email: 'Invalid email format' })
-                }
-            } else {
-                setErrors({ ...errors, email: 'Email is required' })
-            }
-        }
-        ///email validations///
+   
+    // const validate = (input, name) => {
+    //     ///email validations///
+    //     if (name === 'email') {
+    //         if (input.email !== '') {
+    //             // Use a regular expression to check if the input value is a valid email
+    //             const emailPattern = new RegExp(
+    //                 /^([A-Za-z0-9_\-\.])+\@([A-Za-z0-9_\-\.])+\.([A-Za-z]{2,4})$/ //eslint-disable-line
+    //             )
+    //             if (emailPattern.test(input.email)) {
+    //                 setErrors({ ...errors, email: '' })
+    //             } else {
+    //                 setErrors({ ...errors, email: 'Invalid email format' })
+    //             }
+    //         } else {
+    //             setErrors({ ...errors, email: 'Email is required' })
+    //         }
+    //     }
+    //     ///email validations///
 
-        if (name === 'password') {
-            //password length//
-            if (input.password.includes(' ')) {
-                setErrors({
-                    ...errors,
-                    password: 'password cannot have blank spaces',
-                })
-            } else if (
-                input.password.length < 8 ||
-                input.password.length > 12
-            ) {
-                setErrors({
-                    ...errors,
-                    password:
-                        'password must be between 8 and 12 characters long',
-                })
-            } else if (
-                input.password.length >= 8 &&
-                input.password.length <= 12
-            ) {
-                setErrors({ ...errors, password: '' })
-            } else {
-                setErrors({ ...errors, password: 'password is required' })
-            }
-            //password length//
-        }
-    }
+    //     if (name === 'password') {
+    //         //password length//
+    //         if (input.password.includes(' ')) {
+    //             setErrors({
+    //                 ...errors,
+    //                 password: 'password cannot have blank spaces',
+    //             })
+    //         } else if (
+    //             input.password.length < 8 ||
+    //             input.password.length > 12
+    //         ) {
+    //             setErrors({
+    //                 ...errors,
+    //                 password:
+    //                     'password must be between 8 and 12 characters long',
+    //             })
+    //         } else if (
+    //             input.password.length >= 8 &&
+    //             input.password.length <= 12
+    //         ) {
+    //             setErrors({ ...errors, password: '' })
+    //         } else {
+    //             setErrors({ ...errors, password: 'password is required' })
+    //         }
+    //         //password length//
+    //     }
+    // }
 
     ///validations ///
